requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dj-database-url==2.3.0
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
exceptiongroup==1.2.2
fastapi==0.115.12
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==20.1.0
h11==0.14.0
idna==3.10
multidict==6.2.0
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pyserial==3.5
pywin32==306
pyzbar==0.1.9
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.46.2
twilio==9.5.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.1
whitenoise==6.9.0
yarl==1.19.0
zxing==1.0.3
